Refactor all tests
This commit updates the system tests by simplifying the structure of the specs for the blog and pages, improving readability and maintainability. It also modifies the home_spec to better reflect the purpose of the test. Additionally, Capybara's automatic screenshot feature is disabled to streamline test output. These changes enhance the overall clarity and effectiveness of the test suite.